gps trail time override

+++ src/task/import/gps/gpstrailimporttask.cpp
@@ -255,14 +255,14 @@ void GPSTrailImportTask::useTodOffset(bool value)
 
 /**
 */
-float GPSTrailImportTask::todOffset() const
+double GPSTrailImportTask::todOffset() const
 {
     return settings_.tod_offset;
 }
 
 /**
 */
-void GPSTrailImportTask::todOffset(float tod_offset)
+void GPSTrailImportTask::todOffset(double tod_offset)
 {
     loginf << "GPSTrailImportTask: todOffset: value " << tod_offset;
